move copy_template to fileops module
